--- js/get-next-state.js
@@ -1,13 +1,0 @@
-import {Answer} from './data/game-data';
-
-const getNextState = (state, answer) => {
-  const nextState = Object.assign({}, state);
-  nextState.answers.push(answer);
-  nextState.level++;
-  if (answer === Answer.WRONG) {
-    nextState.lives--;
-  }
-  return nextState;
-};
-
-export default getNextState;
